+++ src/ts/Managers/AuctionCampaignManager.ts
@@ -67,7 +67,8 @@ export class AuctionCampaignManager extends CampaignManager {
             for(const placement in placements) {
                 if(placements.hasOwnProperty(placement)) {
                     placementRequest[placement] = {
-                        adTypes: placements[placement].getAdTypes()
+                        adTypes: placements[placement].getAdTypes(),
+                        allowSkip: placements[placement].allowSkip()
                     };
                 }
             }
@@ -103,23 +104,20 @@ export class AuctionCampaignManager extends CampaignManager {
                 }
             }
 
-            let chain = Promise.resolve();
+            const promises: Array<Promise<void>> = [];
 
             for(const placement of noFill) {
-                chain = chain.then(() => {
-                    return this.handlePlcNoFill(placement);
-                });
+                promises.push(this.handlePlcNoFill(placement));
             }
 
             for(const mediaId in fill) {
                 if(fill.hasOwnProperty(mediaId)) {
-                    chain = chain.then(() => {
-                        return this.handlePlcCampaign(fill[mediaId], json.media[mediaId].contentType, json.media[mediaId].content, json.media[mediaId].trackingUrls);
-                    });
+                    promises.push(this.handlePlcCampaign(fill[mediaId], json.media[mediaId].contentType, json.media[mediaId].content, json.media[mediaId].trackingUrls, json.media[mediaId].adType, json.media[mediaId].creativeId, json.media[mediaId].seatId, json.correlationId));
                 }
             }
 
-            return chain.catch(error => {
+            return Promise.all(promises).catch(error => {
+                // todo: catch errors by placement
                 return this.handlePlcError(error);
             });
         } else {
@@ -127,7 +125,7 @@ export class AuctionCampaignManager extends CampaignManager {
         }
     }
 
-    private handlePlcCampaign(placements: string[], contentType: string, content: string, trackingUrls?: { [eventName: string]: string[] }): Promise<void> {
+    private handlePlcCampaign(placements: string[], contentType: string, content: string, trackingUrls?: { [eventName: string]: string[] }, adType?: string, creativeId?: string, seatId?: number, correlationId?: string): Promise<void> {
         const abGroup: number = this._configuration.getAbGroup();
         const gamerId: string = this._configuration.getGamerId();
 
@@ -144,20 +142,20 @@ export class AuctionCampaignManager extends CampaignManager {
                 }
 
             case 'programmatic/vast':
-                return this.parseVastCampaignHelper(content, gamerId, abGroup, trackingUrls).then((vastCampaign) => {
+                return this.parseVastCampaignHelper(content, gamerId, abGroup, trackingUrls, undefined, adType, creativeId, seatId, correlationId).then((vastCampaign) => {
                     return this.setupPlcCampaignAssets(placements, vastCampaign);
                 });
 
             case 'programmatic/mraid-url':
                 const jsonMraidUrl = JsonParser.parse(content);
                 jsonMraidUrl.id = this.getProgrammaticCampaignId();
-                const mraidUrlCampaign = new MRAIDCampaign(jsonMraidUrl, gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : abGroup, jsonMraidUrl.inlinedUrl, undefined, trackingUrls);
+                const mraidUrlCampaign = new MRAIDCampaign(jsonMraidUrl, gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : abGroup, jsonMraidUrl.inlinedUrl, undefined, trackingUrls, adType, creativeId, seatId, correlationId);
                 return this.setupPlcCampaignAssets(placements, mraidUrlCampaign);
 
             case 'programmatic/mraid':
                 const jsonMraid = JsonParser.parse(content);
                 jsonMraid.id = this.getProgrammaticCampaignId();
-                const mraidCampaign = new MRAIDCampaign(jsonMraid, gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : abGroup, undefined, jsonMraid.markup, trackingUrls);
+                const mraidCampaign = new MRAIDCampaign(jsonMraid, gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : abGroup, undefined, jsonMraid.markup, trackingUrls, adType, creativeId, seatId, correlationId);
                 return this.setupPlcCampaignAssets(placements, mraidCampaign);
 
             default:
@@ -167,7 +165,7 @@ export class AuctionCampaignManager extends CampaignManager {
     }
 
     private setupPlcCampaignAssets(placements: string[], campaign: Campaign): Promise<void> {
-        return this._assetManager.setup(campaign, true).then(() => {
+        return this._assetManager.setup(campaign).then(() => {
             for(const placement of placements) {
                 this.onCampaign.trigger(placement, campaign);
             }