+++ src/ts/Test/Unit/Models/MetaData/AdapterMetaDataTest.ts
@@ -15,7 +15,7 @@ class TestStorageApi extends StorageApi {
         this._storage = data;
     }
 
-    public get(storageType: StorageType, key: string): Promise<string | number> {
+    public get<T>(storageType: StorageType, key: string): Promise<T> {
         try {
             switch(key) {
                 case 'adapter.name.value':
@@ -81,13 +81,16 @@ describe('AdapterMetaDataTest', () => {
 
         const metaDataManager: MetaDataManager = new MetaDataManager(nativeBridge);
         return metaDataManager.fetch(AdapterMetaData).then(metaData => {
-            assert.isDefined(metaData, 'AdapterMetaData is not defined');
-            assert.equal(metaData.getName(), 'test_name', 'AdapterMetaData.getName() did not pass through correctly');
-            assert.equal(metaData.getVersion(), 'test_version', 'AdapterMetaData.getVersion() did not pass through correctly');
-            assert.deepEqual(metaData.getDTO(), {
-                adapterName: 'test_name',
-                adapterVersion: 'test_version'
-            }, 'AdapterMetaData.getDTO() produced invalid output');
+            if(metaData) {
+                assert.equal(metaData.getName(), 'test_name', 'AdapterMetaData.getName() did not pass through correctly');
+                assert.equal(metaData.getVersion(), 'test_version', 'AdapterMetaData.getVersion() did not pass through correctly');
+                assert.deepEqual(metaData.getDTO(), {
+                    adapterName: 'test_name',
+                    adapterVersion: 'test_version'
+                }, 'AdapterMetaData.getDTO() produced invalid output');
+            } else {
+                throw new Error('AdapterMetaData is not defined');
+            }
         });
     });
 
@@ -114,9 +117,12 @@ describe('AdapterMetaDataTest', () => {
 
         const metaDataManager: MetaDataManager = new MetaDataManager(nativeBridge);
         return metaDataManager.fetch(AdapterMetaData).then(metaData => {
-            assert.isDefined(metaData, 'AdapterMetaData is not defined');
-            assert.equal(metaData.getName(), 'test_name', 'AdapterMetaData.getName() did not pass through correctly');
-            assert.equal(metaData.getVersion(), undefined, 'AdapterMetaData.getVersion() did not pass through correctly');
+            if(metaData) {
+                assert.equal(metaData.getName(), 'test_name', 'AdapterMetaData.getName() did not pass through correctly');
+                assert.equal(metaData.getVersion(), undefined, 'AdapterMetaData.getVersion() did not pass through correctly');
+            } else {
+                throw new Error('AdapterMetaData is not defined');
+            }
         });
     });
 });