+++ src/ts/Test/Unit/Utilities/DiagnosticsTest.ts
@@ -26,6 +26,10 @@ describe('DiagnosticsTest', () => {
         }, Platform.TEST, false);
     });
 
+    after(() => {
+        HttpKafka.setRequest(undefined);
+    });
+
     it('should not allow primitives as root values', () => {
         const request = new Request(nativeBridge, new WakeUpManager(nativeBridge));
         resolvedPromise = Promise.resolve(TestFixtures.getOkNativeResponse());
@@ -103,7 +107,7 @@ describe('DiagnosticsTest', () => {
 
         return resolvedPromise.then(() => {
             sinon.assert.calledWith(<sinon.SinonStub>request.post,
-                'https://httpkafka.unityads.unity3d.com/v1/events', '{"common":{"client":{"gameId":"12345","testMode":false,"bundleId":"com.unity3d.ads.example","bundleVersion":"2.0.0-test2","sdkVersion":2000,"sdkVersionName":"2.0.0-alpha2","platform":"android","encrypted":false,"configUrl":"http://example.com/config.json","webviewUrl":"http://example.com/index.html","webviewHash":null,"webviewVersion":"2.0.0-webview-test"},"device":null,"country":"fi"}}\n{"type":"ads.sdk2.diagnostics","msg":{"type":"test","test":{"test":true}}}');
+                'https://httpkafka.unityads.unity3d.com/v1/events', '{"common":{"client":{"gameId":"12345","testMode":false,"bundleId":"com.unity3d.ads.example","bundleVersion":"2.0.0-test2","sdkVersion":2000,"sdkVersionName":"2.0.0-alpha2","platform":"android","encrypted":false,"configUrl":"http://example.com/config.json","webviewUrl":"http://example.com/index.html","webviewHash":null,"webviewVersion":"2.0.0-webview-test","initTimestamp":0,"reinitialized":false},"device":null,"country":"fi"}}\n{"type":"ads.sdk2.diagnostics","msg":{"type":"test","test":{"test":true}}}');
         });
     });
 });